Backport Compiler
Backport scala.internal.Compiler from sbt 2.x so we can merge easier.
